refactor(skeleton): wip
