+++ apps/skeleton-app/src/modules/ApplicationRouting/navigation-menu/PanelIdCardSkeleton.tsx
