+++ apps/skeleton-app/src/modules/ApplicationRouting/panel/FullPagePanel.tsx
@@ -3,7 +3,7 @@
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {FunctionComponent} from 'react';
 import {AddPanel, ApplicationMatchingContextWithoutParentTuple} from '../types';
-import {PanelContent} from '../PanelContent';
+import {PanelContent} from './content/PanelContent';
 import {Outlet, useOutletContext} from 'react-router-dom';
 
 interface IFullPagePanelProps {
