+++ apps/skeleton-app/src/modules/ApplicationRouting/panel/content/PanelContent.module.css
